requirements.txt
pandas
numpy
scikit-learn
seaborn
matplotlib
streamlit
sqlite-utils
nltk
joblib
regex
gensim
umap-learn